requirements.txt
asgiref==3.8.1
astroid==3.2.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.2.0
beautifulsoup4==4.12.3
boto3==1.34.119
botocore==1.34.119
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Deprecated==1.2.14
dill==0.3.8
Django==5.0.6
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
lazy-object-proxy==1.10.0
mccabe==0.7.0
msgpack==1.0.8
packaging==24.0
phonenumbers==8.13.38
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycountry==24.6.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==1.7.1
pylint==3.2.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
s3transfer==0.10.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stripe==9.9.0
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.5
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==2.2.1
wget==3.2
wrapt==1.16.0